The H2Truncator now resets the sequences.
The reason this is handy is because now you can deterministically expect
which id gets assigned to a new Entity in your unit tests. This could
cause issues when running the test in isolation vs running the whole
test suite.

For example when testing the following suite you expect both test to pass:

```Java
@Test
public void save() {
    Car car = carRepository.save(new Car);
    assertEquals(car.id, 1);
}

@Test
public void save_withLicencePlate() {
    Car car = carRepository.save(new Car("PB-HL-12"));
    assertEquals(1, car.id); // <- This id will be 2!
    assertEquals("PB-HL-12", car.licensePlate);
}
```

However this is not true at the moment, because the in `save_withLicencePlate`
the cars 'id' table is not reset so it has id 2. This is confusing.

However running only `save_withLicencePlate` will work.

Closes #1