#3 Added the option to add your own tables to exclude from the truncation process. The excluding process is based on AntPathMatcher.match, so you can either pass whole words, or use asterisk to determine either prefix or postfix groups (such as views).
Also refactored the Truncator logic to delegate more to the AbstractDatabaseTruncator class.

Put the logic in place for sequence restarts, but disabled for Postgres. Somehow Postgres still seems to think the number has already been used and throws an error. Not sure what's the cause of this. Perhaps some residual record tracking ID usage?

v0.3.0