Speeded up Postgres truncation by combining queries
Previously, separate queries where executed when
selecting the tables where the row count was larger
than 0 or when performing the truncation/deletion
itself.

Both these queries have been merged into a single
query.

After the change, when running the tests of the application
I was using this library on, the PostgresDeletionStrategy
was actually slightly faster than the PostgresOptimizedDeletionStrategy.
This has however only been tested one time with one single application
so more research is needed to determine if this change makes the optimized
deletion strategy obsolete.

I could not test the PostgresTruncationStrategy as this also did not work
in my application with the previous code.